feat: soundboard with custom thumbnail

+++ src/main/java/com/safjnest/commands/audio/soundboard/SoundboardDelete.java
@@ -37,6 +37,10 @@ public class SoundboardDelete extends SlashCommand{
 	@Override
 	protected void execute(SlashCommandEvent event) {
         String soundboardID = event.getOption("name").getAsString();
+        if(!DatabaseHandler.soundboardExists(soundboardID, event.getGuild().getId(), event.getUser().getId())) {
+            event.deferReply(true).addContent("Soundboard does not exist or you dont have permission to play the selected one.").queue();
+            return;
+        }
         
         DatabaseHandler.deleteSoundboard(soundboardID);
